VALIDATION FILES
Attached is the validation sample file. Download the two images as they are from the validation set, and then download the best path. The only caveat is you can't seem to manually input the path of weights into the run time like you can for images/csvs/etc. So just quickly add it to drive and input the path when prompted in the Notebook